Initial commit
This will be counted as the 0.0.0 release of ssorg. I will continue to improve and debug this program, but this is about as much as I will do for this. This is only intended for personal use and anyone is free to fork, but PRs will most likely be ignored if they don't involve the removal of a bug that is significant.